+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode/
 *.code-workspace
 ./data/
+./idea/
