+++ src/config/config.json
@@ -10,7 +10,7 @@
     "WAL_BUFFER_SIZE": 1024,
     "WAL_SEGMENT_SIZE": 4096,
     "BLOOM_FILTER_FALSE_POSITIVE_RATE": 0.01,
-    "BLOOM_FILTER_EXPECTED_ELEMENTS": 1000,
+    "BLOOM_FILTER_EXPECTED_ELEMENTS": 10,
     "LSM_LEVELS": 2,
     "LSM_BASE_DIR": "data/sstable",
     "COMPACTION_THRESHOLD": 3,
