+++ src/service/ss_compacter/ss_compacter_st.go
@@ -45,7 +45,6 @@ func (sc *SSCompacterST) CheckCompactionConditions(bm *block_manager.BlockManage
 		for len(sstFiles) >= CONFIG.CompactionThreshold && level < CONFIG.LSMLevels {
 			toCompact := sstFiles[:CONFIG.CompactionThreshold]
 			sstFiles = sstFiles[CONFIG.CompactionThreshold:]
-			// You may need to create a FileWriter for the next level here
 			lvlDir := fmt.Sprintf("lvl%d", level+1)
 			fw := file_writer.NewFileWriter(bm, CONFIG.BlockSize, "sstable/"+lvlDir+"/sstable_"+uuid.New().String()+".db")
 			sc.compactTables(toCompact, fw, bm)
@@ -64,13 +63,10 @@ func (sc *SSCompacterST) compactTables(tables []string, fw *file_writer.FileWrit
 	totalItems := 0                                    // total number of items across all tables
 	for i := range tables {
 		counts[i] = int(pool.GetMetadata(i).Getnum_of_items())
-		em:= fmt.Errorf("Error getting metadata for table")
 		totalItems += counts[i]
-		currKeys[i], currValues[i], _, em = pool.ReadNextVal(i) // Read the first key and value from each table
-		fmt.Print(em)
+		currKeys[i], currValues[i], _, _ = pool.ReadNextVal(i) // Read the first key and value from each table
 	}
-	fmt.Print(currKeys)
-
+	// For Index
 	keys := []string{}
 	blockOffsets := []int{}
 	currBlockOffset := -1
@@ -79,7 +75,7 @@ func (sc *SSCompacterST) compactTables(tables []string, fw *file_writer.FileWrit
 	// merkle := merkle_tree.InitializeMerkleTree(totalItems)
 	for !areAllValuesZero(counts) {
 		minIndex := getMinValIndex(currKeys)
-		removeDuplicateKeys(currKeys, minIndex, pool)
+		removeDuplicateKeys(currKeys, minIndex)
 		bloom.Add(currKeys[minIndex])
 		// merkle.AddLeaf(string(keyBytes[minIndex]), valBytes) // Add to Merkle tree
 		fullVal := append(ss_parser.SizeAndValueToBytes(currKeys[minIndex]), ss_parser.SizeAndValueToBytes(currValues[minIndex])...)
@@ -88,16 +84,13 @@ func (sc *SSCompacterST) compactTables(tables []string, fw *file_writer.FileWrit
 			currBlockOffset = newBlockOffset
 			keys = append(keys, currKeys[minIndex])
 			blockOffsets = append(blockOffsets, currBlockOffset)
-
 		}
 		currKeys[minIndex] = "" 
 		updateValsAndCounts(currKeys, currValues, counts, pool)
 	}
-
 	fw.Write(nil, true, nil) // Write end of file marker
 	summaryKeys, summaryOffsets := ss_parser.SerializeIndexGetOffsets(keys, blockOffsets, fw) // Write index offsets
 	initialSummaryOffset := fw.Write(nil, true, nil)
-
 	ss_parser.SerializeSummary(summaryKeys, summaryOffsets, fw)
 	prefixFilter := bloom_filter.NewPrefixBloomFilter()
